backend/resume-builder/requirements.txt
fastapi
uvicorn
jinja2
weasyprint
langchain 
langchain-google-genai
google-generativeai
python-dotenv
motor
pymongo
PyJWT